Re-arrange linux tests into some separate files for each target.
